[Enhancement] refactor and optimize iceberg [stage 3]
