+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Utils.java
@@ -23,8 +23,6 @@ import com.starrocks.catalog.OlapTable;
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Table;
 import com.starrocks.catalog.Type;
-import com.starrocks.connector.iceberg.ScalarOperatorToIcebergExpr;
-import com.starrocks.connector.iceberg.cost.IcebergTableStatisticCalculator;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.optimizer.operator.Operator;
@@ -43,8 +41,6 @@ import com.starrocks.sql.optimizer.operator.scalar.ConstantOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
 import com.starrocks.sql.optimizer.statistics.ColumnStatistic;
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.iceberg.expressions.Expression;
-import org.apache.iceberg.types.Types;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -368,15 +364,8 @@ public class Utils {
             } else if (operator instanceof LogicalIcebergScanOperator) {
                 IcebergTable table = (IcebergTable) scanOperator.getTable();
                 try {
-                    List<ScalarOperator> predicates = Utils.extractConjuncts(operator.getPredicate());
-                    Types.StructType schema = table.getIcebergTable().schema().asStruct();
-                    ScalarOperatorToIcebergExpr.IcebergContext icebergContext =
-                            new ScalarOperatorToIcebergExpr.IcebergContext(schema);
-                    Expression icebergPredicate = new ScalarOperatorToIcebergExpr().convert(predicates, icebergContext);
-                    List<ColumnStatistic> columnStatisticList = IcebergTableStatisticCalculator.getColumnStatistics(
-                            icebergPredicate, table.getIcebergTable(),
-                            scanOperator.getColRefToColumnMetaMap());
-                    return columnStatisticList.stream().anyMatch(ColumnStatistic::isUnknown);
+                    // iceberg metadata doesn't have ndv, we default to unknown for all iceberg table column statistics.
+                    return true;
                 } catch (Exception e) {
                     LOG.warn("Iceberg table {} get column failed. error : {}", table.getName(), e);
                     return true;
